scorer/requirements.txt
fastapi
uvicorn
lightgbm
scikit-learn==1.3.2
pandas
numpy
joblib
psycopg2-binary
requests
arm-pyart
